Relocate everything modwinder-related
