+++ src/main/java/ladylib/client/shader/BaseShaders.java
@@ -1,5 +1,6 @@
 package ladylib.client.shader;
 
+import ladylib.modwinder.client.gui.MWShaders;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
@@ -15,12 +16,11 @@ public final class BaseShaders {
     public static final ResourceLocation BASE_FRAGMENT = new ResourceLocation(MOD_ID, SHADER_LOCATION_PREFIX + "fragment_base.fsh");
 
     public static final ResourceLocation GREYSCALE = new ResourceLocation(MOD_ID, "greyscale");
-    public static final ResourceLocation ROUNDISH = new ResourceLocation(MOD_ID, "roundish");
 
     @SubscribeEvent
     public static void onShaderRegistry(ShaderRegistryEvent event) {
         event.registerFragmentShader(GREYSCALE);
-        event.registerFragmentShader(ROUNDISH);
+        event.registerFragmentShader(MWShaders.ROUNDISH);
     }
 
     private BaseShaders() {
