+++ src/main/java/ladylib/modwinder/installer/InstallationException.java
@@ -1,4 +1,4 @@
-package ladylib.installer;
+package ladylib.modwinder.installer;
 
 public class InstallationException extends RuntimeException {
     public InstallationException(String message, Throwable cause) {
